requirements.txt
setuptools >= 80.0
aiosqlite==0.21.0
dateparser==1.2.2
numpy==2.3.3
python-dotenv==1.1.1
python-telegram-bot[job-queue]==22.5
Requests==2.32.5
skia_python==138.0
urllib3==2.5.0
